made a util folder
